Remove some allows

+++ chains/container-chains/runtime-templates/simple/src/lib.rs
@@ -1177,8 +1177,7 @@ impl_runtime_apis! {
                         );
                     });
 
-                    #[allow(clippy::cast_sign_loss)]
-                    Some((assets, fee_index as u32, dest, verify))
+                    Some((assets, fee_index, dest, verify))
                 }
             }
